requirements.txt
streamlit==1.32.0
pandas==2.1.4
numpy==1.26.3
pillow==10.2.0
plotly==5.18.0
matplotlib==3.8.2
scikit-learn==1.4.0
python-dotenv==1.0.0
openai==1.13.3 